Games Controller arreglado y Score tipo Int

+++ InaDraft_API/InaDraft_DAL/Lists/clsGameListDAL.cs
@@ -39,7 +39,7 @@ namespace InaDraft_DAL.Lists
         {
             return new clsGame((int)MyReader["Id"],
                                         (string)MyReader["UserNick"],
-                                        (double)MyReader["Score"],
+                                        (int)MyReader["Score"],
                                         (DateTime)MyReader["Date"], 
                                         (int)MyReader["FormationId"]);
         }
